Update to rapier 0.27.0-beta.0
