+++ src/pipeline/mod.rs
@@ -2,12 +2,10 @@ pub use self::debug_render_pipeline::*;
 pub use self::event_queue::*;
 pub use self::physics_hooks::*;
 pub use self::physics_pipeline::*;
-pub use self::query_pipeline::*;
 pub use self::serialization_pipeline::*;
 
 mod debug_render_pipeline;
 mod event_queue;
 mod physics_hooks;
 mod physics_pipeline;
-mod query_pipeline;
 mod serialization_pipeline;
